주요 하나 남음

+++ src/main/java/com/green/airline/dto/UseMileageDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class UseMileageDto {
 
 	private Date useDate;
-	private Long useMileage;
+	private int useMileage;
 	private String description;
 	private int productPrice;
 	private int hiddenCount;
